requirements.txt
mitmproxy~=10.2.4
pystray~=0.19.5
pillow~=10.3.0